renamed unit to size and used walmart api to gather information on products

+++ src/Application/ProductStocks/Queries/ProductStockDto.cs
@@ -9,5 +9,5 @@ public class ProductStockDto : IMapFrom<ProductStock>
 
     public string Name { get; set; }
 
-    public int Units { get; set; }
+    public float Units { get; set; }
 }
